pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcam-classification"
version = "1.0.0"
description = "FDA-compliant center-aware deep learning for PCam histopathology tumor detection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "PCam Classification Team"}
]
keywords = [
    "deep-learning",
    "medical-imaging",
    "histopathology",
    "cancer-detection",
    "pytorch",
    "fda-compliance"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "h5py>=3.7.0",
    "Pillow>=9.0.0",
    "opencv-python>=4.5.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tqdm>=4.64.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "timm>=0.9.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6",
    "prometheus-client>=0.17.0",
]
mlops = [
    "mlflow>=2.5.0",
    "optuna>=3.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
    "pre-commit>=3.3.0",
]
all = [
    "pcam-classification[api,mlops,dev]",
]

[project.urls]
Homepage = "https://github.com/your-org/cancer-classification"
Documentation = "https://github.com/your-org/cancer-classification#readme"
Repository = "https://github.com/your-org/cancer-classification"
Issues = "https://github.com/your-org/cancer-classification/issues"

[project.scripts]
pcam-train = "scripts.train:main"
pcam-evaluate = "scripts.evaluate_model:main"
pcam-export = "scripts.export_models:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | data
    | experiments
    | mlruns
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "build", "dist", "data", "experiments", "mlruns"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["data", "experiments", "mlruns"]
